refactor(docs-infra): move windows-chromium-path to adev
Move windows-chromium-path help from aio to adev
